move Kafka functions to kafka_utils.py add support for addition sections in Kafka configuration file
