+++ nslsii/tests/test_kafka_configuration.py
@@ -1,23 +1,24 @@
 import os
-from pathlib import Path
 
 import pytest
 
-from nslsii import (
-    configure_kafka_publisher,
+from nslsii import configure_kafka_publisher
+
+from nslsii.kafka_utils import (
     _read_bluesky_kafka_config_file,
-    _SubscribeKafkaPublisherDetails,
-    _SubscribeKafkaQueueThreadPublisherDetails
 )
 
+# these test configurations include localhost:9092
+# because configure_kafka_publisher verifies that
+# a connection can be made to a broker
 
 test_bluesky_kafka_config_true = """\
 ---
   abort_run_on_kafka_exception: true
   bootstrap_servers:
+    - localhost:9092
     - kafka1:9092
     - kafka2:9092
-    - kafka3:9092
   runengine_producer_config:
     acks: 0
     message.timeout.ms: 3000
@@ -28,29 +29,59 @@ test_bluesky_kafka_config_false = """\
 ---
   abort_run_on_kafka_exception: false
   bootstrap_servers:
+    - localhost:9092
     - kafka1:9092
     - kafka2:9092
-    - kafka3:9092
   runengine_producer_config:
     acks: 0
     message.timeout.ms: 3000
     compression.codec: snappy
 """
 
+test_bluesky_kafka_config_security_section = """\
+---
+  abort_run_on_kafka_exception: false
+  bootstrap_servers:
+    - localhost:9092
+    - kafka2:9092
+    - kafka3:9092
+  producer_consumer_security_config:
+    security.protocol: SASL_SSL
+    sasl.mechanisms: PLAIN
+    ssl.ca.location: /etc/ssl/certs/ca-bundle.crt
+  consumer_config:
+    auto.offset.reset: latest
+  runengine_producer_config:
+    compression.codec: snappy
+    security.protocol: SASL_SSL
+    sasl.mechanisms: PLAIN
+    ssl.ca.location: /etc/ssl/certs/ca-bundle.crt
+  runengine_topics:
+    - "{endstation}.bluesky.runengine.documents"
+    - "{endstation}.bluesky.runengine.{document_name}.documents"
+"""
 
-def test_bluesky_kafka_config_path_env_var(tmp_path, RE):
+
+def test_bluesky_kafka_config_path_env_var(tmp_path, RE, temporary_topics):
     """Test specifying a configuration file path by environment variable."""
-    # write a temporary file for this test
-    test_config_file_path = tmp_path / "bluesky_kafka_config_content.yml"
-    with open(test_config_file_path, "wt") as f:
-        f.write(test_bluesky_kafka_config_false)
-        # add an extra item to test for later
-        f.write(f"  config_file_path: {test_config_file_path}")
+    with temporary_topics(topics=["abc.bluesky.runengine.documents"]) as (
+        beamline_topic,
+    ):
+        # write a temporary file for this test
+        test_config_file_path = tmp_path / "bluesky_kafka_config_content.yml"
+        with open(test_config_file_path, "wt") as f:
+            f.write(test_bluesky_kafka_config_false)
+            # add an extra item to test for later
+            f.write(f"  config_file_path: {test_config_file_path}")
 
-    os.environ["BLUESKY_KAFKA_CONFIG_PATH"] = str(test_config_file_path)
-    bluesky_kafka_configuration, publisher_details = configure_kafka_publisher(RE, "abc")
+        os.environ["BLUESKY_KAFKA_CONFIG_PATH"] = str(test_config_file_path)
+        bluesky_kafka_configuration, publisher_details = configure_kafka_publisher(
+            RE, "abc"
+        )
 
-    assert bluesky_kafka_configuration["config_file_path"] == str(test_config_file_path)
+        assert bluesky_kafka_configuration["config_file_path"] == str(
+            test_config_file_path
+        )
 
 
 def test_bluesky_kafka_config_path_env_var_negative(tmp_path, RE):
@@ -85,9 +116,9 @@ def test__read_bluesky_kafka_config_file(tmp_path):
     bluesky_kafka_config = _read_bluesky_kafka_config_file(str(test_config_file_path))
 
     assert bluesky_kafka_config["bootstrap_servers"] == [
+        "localhost:9092",
         "kafka1:9092",
         "kafka2:9092",
-        "kafka3:9092",
     ]
 
     runengine_producer_config = bluesky_kafka_config["runengine_producer_config"]
@@ -95,6 +126,59 @@ def test__read_bluesky_kafka_config_file(tmp_path):
     assert runengine_producer_config["acks"] == 0
     assert runengine_producer_config["message.timeout.ms"] == 3000
     assert runengine_producer_config["compression.codec"] == "snappy"
+
+
+def test__read_bluesky_kafka_config_file_producer_consumer_security(tmp_path):
+    # write a temporary file for this test
+    test_config_file_path = tmp_path / "bluesky_kafka_config_content.yml"
+    with open(test_config_file_path, "wt") as f:
+        f.write(test_bluesky_kafka_config_security_section)
+
+    bluesky_kafka_config = _read_bluesky_kafka_config_file(str(test_config_file_path))
+
+    assert bluesky_kafka_config["bootstrap_servers"] == [
+        "localhost:9092",
+        "kafka2:9092",
+        "kafka3:9092",
+    ]
+
+    producer_consumer_security_config = bluesky_kafka_config[
+        "producer_consumer_security_config"
+    ]
+    assert len(producer_consumer_security_config) == 3
+    assert producer_consumer_security_config["security.protocol"] == "SASL_SSL"
+    assert producer_consumer_security_config["sasl.mechanisms"] == "PLAIN"
+    assert (
+        producer_consumer_security_config["ssl.ca.location"]
+        == "/etc/ssl/certs/ca-bundle.crt"
+    )
+
+    runengine_producer_config = bluesky_kafka_config["runengine_producer_config"]
+    assert len(runengine_producer_config) == 4
+    assert runengine_producer_config["compression.codec"] == "snappy"
+    assert producer_consumer_security_config["security.protocol"] == "SASL_SSL"
+    assert producer_consumer_security_config["sasl.mechanisms"] == "PLAIN"
+    assert (
+        producer_consumer_security_config["ssl.ca.location"]
+        == "/etc/ssl/certs/ca-bundle.crt"
+    )
+
+
+def test__read_bluesky_kafka_config_file_runengine_topics(tmp_path):
+    # write a temporary file for this test
+    test_config_file_path = tmp_path / "bluesky_kafka_config_content.yml"
+    with open(test_config_file_path, "wt") as f:
+        f.write(test_bluesky_kafka_config_security_section)
+
+    bluesky_kafka_config = _read_bluesky_kafka_config_file(str(test_config_file_path))
+
+    runengine_topics = bluesky_kafka_config["runengine_topics"]
+    assert len(runengine_topics) == 2
+    assert runengine_topics[0] == "{endstation}.bluesky.runengine.documents"
+    assert (
+        runengine_topics[1]
+        == "{endstation}.bluesky.runengine.{document_name}.documents"
+    )
 
 
 def test__read_bluesky_kafka_config_file_failure(tmp_path):
@@ -136,14 +220,14 @@ def test_configure_kafka_publisher_abort_run_true(tmp_path, RE):
         f.write(test_bluesky_kafka_config_true)
 
     bluesky_kafka_configuration, publisher_details = configure_kafka_publisher(
-        RE,
-        "abc",
-        override_config_path=test_config_file_path
+        RE, "abc", override_config_path=test_config_file_path
     )
 
     assert publisher_details.__class__.__name__ == "SubscribeKafkaPublisherDetails"
     assert publisher_details.beamline_topic == "abc.bluesky.runengine.documents"
-    assert publisher_details.bootstrap_servers == "kafka1:9092,kafka2:9092,kafka3:9092"
+    assert (
+        publisher_details.bootstrap_servers == "localhost:9092,kafka1:9092,kafka2:9092"
+    )
     assert publisher_details.re_subscribe_token == 0
 
 
@@ -157,12 +241,15 @@ def test_configure_kafka_publisher_abort_run_false(tmp_path, RE):
         f.write(test_bluesky_kafka_config_false)
 
     bluesky_kafka_configuration, publisher_details = configure_kafka_publisher(
-        RE,
-        "abc",
-        override_config_path=test_config_file_path
+        RE, "abc", override_config_path=test_config_file_path
     )
 
-    assert publisher_details.__class__.__name__ == "SubscribeKafkaQueueThreadPublisherDetails"
+    assert (
+        publisher_details.__class__.__name__
+        == "SubscribeKafkaQueueThreadPublisherDetails"
+    )
     assert publisher_details.beamline_topic == "abc.bluesky.runengine.documents"
-    assert publisher_details.bootstrap_servers == "kafka1:9092,kafka2:9092,kafka3:9092"
+    assert (
+        publisher_details.bootstrap_servers == "localhost:9092,kafka1:9092,kafka2:9092"
+    )
     assert publisher_details.re_subscribe_token is None
